Get rid of the old withEntitiesContainer HOC
Move withEntitiesContainer2.js into withEntitiesContainer.js and update
all imports to use the new path.
