+++ gsa/CMakeLists.txt
@@ -350,7 +350,6 @@ set (GSA_JS_SRC_FILES
      ${GSA_SRC_DIR}/src/web/entities/table.js
      ${GSA_SRC_DIR}/src/web/entities/tagsdialog.js
      ${GSA_SRC_DIR}/src/web/entities/withEntitiesContainer.js
-     ${GSA_SRC_DIR}/src/web/entities/withEntitiesContainer2.js
      ${GSA_SRC_DIR}/src/web/entities/withRowDetails.js
      ${GSA_SRC_DIR}/src/web/entity/block.js
      ${GSA_SRC_DIR}/src/web/entity/box.js
