+++ gsa/src/web/entities/withEntitiesContainer.js
@@ -20,30 +20,71 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
  */
-
 import React from 'react';
 
-import EntitiesContainer from './container.js';
+import {connect} from 'react-redux';
 
-import CacheProvider from '../components/provider/cacheprovider.js';
-import SubscriptionProvider from '../components/provider/subscriptionprovider';
+import getPage from 'web/store/pages/selectors';
 
-const withEntitiesContainer = (gmpname, options = {}) => component => {
-  const EntitiesContainerWrapper = props => (
-    <CacheProvider name={gmpname}>
-      <SubscriptionProvider>
-        {({notify}) => (
-          <EntitiesContainer
-            {...options}
-            {...props}
-            notify={notify}
-            gmpname={gmpname}
-            component={component}
-          />
-        )}
-      </SubscriptionProvider>
-    </CacheProvider>
+import SubscriptionProvider from 'web/components/provider/subscriptionprovider';
+
+import compose from 'web/utils/compose';
+import withGmp from 'web/utils/withGmp';
+import withDownload from 'web/components/form/withDownload';
+import withDialogNotification from 'web/components/notification/withDialogNotifiaction'; // eslint-disable-line max-len
+
+import EntitiesContainer from './container2';
+import {pageFilter} from 'web/store/pages/actions';
+import {isDefined} from 'gmp/utils/identity';
+
+const withEntitiesContainer = (gmpname, {
+  entitiesSelector,
+  loadEntities,
+}) => Component => {
+  let EntitiesContainerWrapper = props => (
+    <SubscriptionProvider>
+      {({notify}) => (
+        <EntitiesContainer
+          {...props}
+          notify={notify}
+          gmpname={gmpname}
+        >
+          {pageProps => (
+            <Component
+              {...pageProps}
+            />
+          )}
+        </EntitiesContainer>
+      )}
+    </SubscriptionProvider>
   );
+
+  const mapStateToProps = (state, props) => {
+    const eSelector = entitiesSelector(state);
+    const pSelector = getPage(state);
+    const filter = pSelector.getFilter(gmpname);
+    const entities = eSelector.getEntities(filter);
+    return {
+      entities,
+      entitiesCounts: eSelector.getEntitiesCounts(filter),
+      filter,
+      isLoading: !isDefined(entities) || eSelector.isLoadingEntities(filter),
+      loadedFilter: eSelector.getLoadedFilter(filter),
+    };
+  };
+
+  const mapDispatchToProps = (dispatch, {gmp}) => ({
+    loadEntities: filter => dispatch(loadEntities({gmp, filter})),
+    updateFilter: filter => dispatch(pageFilter(gmpname, filter)),
+  });
+
+  EntitiesContainerWrapper = compose(
+    withDialogNotification,
+    withDownload,
+    withGmp,
+    connect(mapStateToProps, mapDispatchToProps),
+  )(EntitiesContainerWrapper);
+
   return EntitiesContainerWrapper;
 };
