+++ gsa/src/web/pages/operatingsystems/listpage.js
@@ -32,7 +32,7 @@ import PropTypes from 'web/utils/proptypes';
 import Layout from 'web/components/layout/layout';
 
 import EntitiesPage from 'web/entities/page';
-import withEntitiesContainer from 'web/entities/withEntitiesContainer2';
+import withEntitiesContainer from 'web/entities/withEntitiesContainer';
 
 import DashboardControls from 'web/components/dashboard/controls';
