+++ gsa/src/web/pages/tasks/listpage.js
@@ -35,7 +35,7 @@ import {
 } from 'web/store/entities/tasks';
 
 import EntitiesPage from 'web/entities/page';
-import withEntitiesContainer from 'web/entities/withEntitiesContainer2';
+import withEntitiesContainer from 'web/entities/withEntitiesContainer';
 
 import DashboardControls from 'web/components/dashboard/controls';
